% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_assign.R
\name{assign}
\alias{assign}
\title{Assigns data points to k clusters}
\usage{
assign(X, centers)
}
\arguments{
\item{X}{array Data points of dimension (n,d)}

\item{centers}{array Coordinates of cluster centers, dimension (k,d)}
}
\value{
array Cluster assignments for each point in X, dimension (n,1)
}
\description{
Assigns data points to k clusters
}
\examples{
X = rbind(c(0,0), c(1,1))
centers = rbind(c(0.1, 0.1), c(1.1, 1.1))
assign(X, centers)
}
