% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_elbow.R
\name{avg_sil_score}
\alias{avg_sil_score}
\title{avg_sil_score
This function takes in a vector of cluster labels and a corresponding array of points and calculates the average silhouette score across all clusters.
It returns the value of the average silhouette score.}
\usage{
avg_sil_score(labels, X)
}
\arguments{
\item{labels}{vector: A vector of cluster labels.}

\item{X}{matrix: A (n_points x n_features) matrix of points.}
}
\value{
numeric: The average silhouette score.
}
\description{
avg_sil_score
This function takes in a vector of cluster labels and a corresponding array of points and calculates the average silhouette score across all clusters.
It returns the value of the average silhouette score.
}
\examples{
helper_data <- array(c(c(0, 10, 10), c(0, 10, 11)), dim = c(3,2))
helper_clusters <- c(1, 2, 2)
avg_sil_score(helper_clusters, helper_data)
}
