% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_assign.R
\name{calc_centers}
\alias{calc_centers}
\title{Calculates center coordinates of each cluster}
\usage{
calc_centers(X, centers, labels)
}
\arguments{
\item{X}{array Data points of dimension (n,d)}

\item{centers}{array Coordinates of previous cluster centers, dimension (k,d) (used only to determine number of centers k)}

\item{labels}{array Cluster assignments for each point in X, dimension (n,1)}
}
\value{
array Center coordinate for each cluster, dimension (k,d)
}
\description{
Calculates center coordinates of each cluster
}
\examples{
X = rbind(c(0,0), c(1,0), c(10,0), c(11,0))
centers = rbind(c(0, 0), c(10, 10))
labels = c(1, 1, 2, 2)
calc_centers(X, centers, labels)
}
