% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_assign.R
\name{fit_assign}
\alias{fit_assign}
\title{Finds k clusters in data points and assigns each point to a cluster.}
\usage{
fit_assign(X, k)
}
\arguments{
\item{X}{array: Data points of dimension (n,d)}

\item{k}{int: The number of desired clusters}
}
\value{
list: array of cluster centers dimension(n,d), array of cluster labels dimension (n,1)
}
\description{
Finds k clusters in data points and assigns each point to a cluster.
}
\examples{
X = rbind(c(0,0), c(1,1))
km <- fit_assign(X, 2)
}
