% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_assign.R
\name{init_centers}
\alias{init_centers}
\title{Chooses initial cluster locations using Kmeans++}
\usage{
init_centers(X, k)
}
\arguments{
\item{X}{array Data points of dimension (n,d)}

\item{k}{int The number of desired clusters}
}
\value{
array Initial coordinates of clusters
}
\description{
Chooses initial cluster locations using Kmeans++
}
\examples{
X = rbind(c(0,0), c(1,1))
init_centers(X, 2)
}
