% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_assign.R, R/make_blobs.R
\name{make_blobs}
\alias{make_blobs}
\title{Makes clusters of n points in d dimensions}
\usage{
make_blobs(n, k, d)

make_blobs(n, k, d)
}
\arguments{
\item{n}{int Number of points}

\item{k}{int Number of clusters}

\item{d}{int Number of dimensions}
}
\value{
Dataframe of points

Matrix of points
}
\description{
Makes clusters of n points in d dimensions

Makes clusters of n points in d dimensions
}
\examples{
df = make_blobs(n, k, d)

df = make_blobs(n, k, d)

}
