% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_assign.R
\name{measure_dist}
\alias{measure_dist}
\title{Measures distance from data points to cluster centers}
\usage{
measure_dist(X, centers)
}
\arguments{
\item{X}{array Data points of dimension (n,d)}

\item{centers}{array Coordinates of cluster centers, dimension (k,d)}
}
\value{
array Distance from each point to each center, dimension (n,k)
}
\description{
Measures distance from data points to cluster centers
}
\examples{
X = rbind(c(0,0), c(1,1))
centers = rbind(c(0.1, 0.1), c(1.1, 1.1))
measure_dist(X, centers)
}
