% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_blobs.R
\name{plot_blobs}
\alias{plot_blobs}
\title{Makes scatter plot of points in X}
\usage{
plot_blobs(X, centers, labels = none, title = "")
}
\arguments{
\item{X}{Array (2-dimensional) of points}

\item{centers}{Array (2-dimensional) of cluster centers}
}
\value{
ggplot object
}
\description{
Makes scatter plot of points in X
}
