% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_clusters.R
\name{show_clusters}
\alias{show_clusters}
\title{This functionreduces a data set to 2 dimensions using principle component analysis (PCA) and colours clusters of points.}
\usage{
show_clusters(X, clusters, centroids = None)
}
\arguments{
\item{X}{DataFrame Unprocessed data}

\item{clusters}{vector corresponding cluster for X}

\item{centroids}{array Center coordinate for each cluster}
}
\value{
plot A 2d principle components scatter plot coloured by cluster
}
\description{
This functionreduces a data set to 2 dimensions using principle component analysis (PCA) and colours clusters of points.
}
\examples{
 X = rbind(c(0,0), c(1,0), c(10,0), c(11,0))
centers = rbind(c(0, 0), c(10, 10))
}
